Introduce our own `ModuleT` trait
The reasoning behind this is that the Candle `ModuleT` trait is less
flexible, since it uses `candle_core::Error`. This has two downsides:

- Our own module implementations can only use Candle errors, which may
  not be sufficient to express the errors that can occur.
- It makes it harder to mix errors in a call chain, because we often end
  up with a mixture of methods that return `BoxedError` and
  `candle_core::Error`.

Initially I had a version where the error type was an associated type of
`ModuleT`. However, this put the burden on the user of the trait to
specify the error type everywhere, which is `BoxedError` in most cases
anyway.